fix server.js
