+++ routes/elements.route.js
@@ -6,13 +6,13 @@ var router = express.Router();
 var elemCtrl = require('../controllers/element.controller');
 
 /* Routes that can be accessed only by authenticated users*/
-router.get('/api/elements', elemCtrl.getAll); //get all elements
-router.get('/api/element/:id', elemCtrl.getOne); //get one element
-router.post('/api/element/', elemCtrl.create); //create element
+router.get('/api/elements', elemCtrl.getAllElements); //get all elements
+router.get('/api/element/:id', elemCtrl.getOneElement); //get one element
+router.post('/api/element/', elemCtrl.addElement); //create element
 
 /* Routes that can be accessed only by authenticated & authorized users*/
-router.put('/api/owner/element/:id', elemCtrl.update); //update element
-router.delete('/api/owner/element/:id', elemCtrl.delete); //delete element
+router.put('/api/owner/element/:id', elemCtrl.updateElement); //update element
+router.delete('/api/owner/element/:id', elemCtrl.deleteElement); //delete element
 // User needs to be owner to modify or delete elements
 
 module.exports = router;
